added

+++ 0x00-hello_world/6-size.c
@@ -10,11 +10,11 @@ int i;
 long int il;
 long long int ill;
 float a;
-printf("size of char: %lu byte(s)\n", sizeof(ab));
-printf("size of int: %lu byte(s)\n", sizeof(i));
-printf("size of long int: %lu byte(s)\n", sizeof(il));
-printf("size of long long int: %lu byte(s)\n", sizeof(ill));
-printf("size of float: %lu byte(s)\n", sizeof(a));
+printf("size of char: %lu byte(s)\n", (unsigned long) sizeof(ab));
+printf("size of int: %lu byte(s)\n", (unsigned long)  sizeof(i));
+printf("size of long int: %lu byte(s)\n", (unsigned long) sizeof(il));
+printf("size of long long int: %lu byte(s)\n", (unsigned long) sizeof(ill));
+printf("size of float: %lu byte(s)\n", (unsigned long) sizeof(a));
 
 return (0);
 }
